change some short int to int
git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@3096 eb38b3b0-33ea-0310-94ad-d43fea0d214f

+++ NEGF/Headers/LCR.h
@@ -42,12 +42,12 @@ struct NON_LINEAR_THREE_PART
  short int NY_GRID;
  short int NZ_GRID;
 
- short int num_states;
- short int num_ions;
- short int state_begin;
- short int state_middle;
- short int state_end;
- short int ion_begin;
+ int num_states;
+ int num_ions;
+ int state_begin;
+ int state_middle;
+ int state_end;
+ int ion_begin;
  char name[MAX_PATH];
  char lead_name[MAX_PATH];
